Fix select logic to only "ready" fds with appropriate operations.
POSIX select removes from the given fd sets all fds that are *not*
ready for that particular operation. That was missing in our
logic, so *all* fds selected on were put into *all* ready lists.

This change passes a set of interest operations into fdIsSet,
allowing us to detect whether the fd is actually ready for the
requested operations.

For #2102.