Restructure argv into a single command when doing chdir.
JRuby emulates current dir to be a good JVM citizen, since doing
chdir could really mess up other apps/threads running in the same
process. Unfortunately, posix_spawn does not support chdir, so we
restructure the incoming command to use sh with a chdir prepended
to the script. This worked fine for cases where argv only
contained the command to run, but we missed logic to deal with
additional args.

This commit checks for chdir and argc > 1, building a new command
line string by joining the individual argv elements with ' '.

Fix was to get the update_rubygems command working, which does a
Dir.chdir followed by a Kernel#system.