associate the MethodHandles with classloader which loaded jruby - fixes #2091
the static MethodHandles#publicLookup does not work when the JVM has two (or more) classloaders
where jruby is running in - the lookup uses Object.class as "reference". to separate the jitted methods the lookup needs to be on a per
classloader level. since Ruby.class is different when coming from different classloaders it is used
to separate the lookup of MethodHandles.