Avoid CNFE by setting classloader resource of service .class.
When loading a .jar file, JRuby will also attempt to load a class
named *Service based on the require path of the jar, in order to
boot the extension the jar contains (if present). However this
usually fails, since requiring a jar is also the preferred way to
make a jar's contents available. As a result, every jar loaded by
a JRuby application triggers the *Service search, which usually
results in a ClassNotFoundException. Exceptions are expensive, so
this commit tries to actively look for the *Service.class rather
than falling into the heavier failure logic.