not using lib/jruby.jar to test $CLASSPATH
adding lib/jruby.jar on the fly to Thread.currentThread.contextClassLoader (CCL)
where is parent has already lib/jruby.jar is not an good idea especially
if the CCL is the JRubyClassLoader looking first into itself before going to
its parent CL to load classes.

further I added test to ensure that the added $CLASSPATH entry indeed
is part of the classloader.

fixes #2276