Remove native bits of delegate lib.
We implemented portions of the delegate library as a JRuby ext,
in order to improve performance and avoid the heavy method_missing
logic that used to be in the pure-Ruby version. However, over the
years, the delegate library has improved and no longer has the
same performance characteristics as it did in 1.8 and earlier. In
addition, there have been functionality changes to the pure-Ruby
code that we would have to maintain in the extension.

Therefore, we have opted to return to just the pure-Ruby delegate
library. We will reexamine performance and work with ruby-core to
improve the pure-Ruby version before we consider returning to the
native version.

Fixes #1370.