[Truffle] Fix null FrameSlot when kwarg is not used
When a keyword argument is not used in the method body there is no frame
slot for the argument resulting a NullPointerException when writing the
argument to the frame.

This fix changes the AST translator to create the frame slot if it
cannot be found by using `findOrAddFrameSlot` instead of
`findFrameSlot`.