Skip block-to-proc conversion for explicit block syntax when possible
If an explicit block is captured as a proc, but its only usage is
converting it back to a block and passing it through, we can skip the
conversion entirely, and just pass through the original block.

This pass looks for all explicit block conversions, and checks if
they're ever used in any context besides being passed as a block. If
not, we rewrite the `ReifyProcInstr` to just copy the block, instead.

Given the following Ruby code:

```ruby
def foo(&block)
  bar(&block)
end
```

The instruction sequence before was:

```
%self = recv_self
check_arity req: 0, opt: 0, *r: -1, kw: false, **r: -1
%v_1 = load_implicit_closure
%t_block_2 = reify_closure(%v_1)
line_num(1)
%v_1 = call(FUNCTIONAL, bar, %self, [], &%t_block_2)
return(%v_1)
```

With this patch, the instruction sequence changes to:

```
%self = recv_self
check_arity req: 0, opt: 0, *r: -1, kw: false, **r: -1
%v_1 = load_implicit_closure
%t_block_2 = copy(%v_1)
line_num(1)
%v_1 = call(FUNCTIONAL, bar, %self, [], &%t_block_2)
return(%v_1)
```

Which gives a healthy boost to performance on any code which passes
blocks through explicitly.