Fix #2514: Get rid of broken sanity test for exceptions
I had added that test originally as a sanity test to make sure the
new exception handling code that sets $! and saves/restores that
global variable wasn't broken in 9K (which is different from 1.7)

However, as reported in #2514, there are real scenarios where the
objects will not line up -- specifically because of the proxy wrapping
for java exceptions.

Given that we haven't had any real bug reports from that test all
this time we hvae run tests and apps, and we only got a false positive,
I'm going to get rid of the test and be done with it.