[Truffle] Moved Kernel#print to Rubinius.
To make this work I needed to move the writing to STDOUT elsewhere.  IO#write is the natural place for this to go, although it shouldn't be couple to STDOUT like it currently is.  The implementation is wrong, as was the old one, but wrong in a different way.  The big difference now is /STDOUT can be redefined and Kernel#print will react accordingly.