First pass attempting to optimize startup time
This patch is a bunch of hacky fixes as a proof of
concept attempt to improve startup time.
* This fixes the interpreter to interpret IR instructions
  that the IR builder generates without having to build
  a CFG or run any passes on them. There is no cloning
  of instructions either.
* If the scope runs at least N times (N = 20 right now),
  the CFG is then built, some basic passes are run and
  future runs will use the new optimized scope.

In interpret mode (-X-C), this patch:
* reduces runtime for gem list by 20+% (which manifests as
  improved startup time).
* reduces total user time by ~13% across a bunch of benchmarks.
* reduces wall clock time by ~5% across the same benchmarks.
* improves peak perf on some of those benchmarks and
  degrades peak perf on some.

In default mode, this patch:
* reduces total user time by ~20% across a bunch of benchmarks.
* reduces wall clock time by ~11% across the same benchmarks.

This was passing all tests in an earlier version of this
patch before I broke something messing and tweaking
something. It is late at night, so pushing this as is.

This branch is not ready for merge. It requires:
* code cleanup
* get rid of hacks
* more testing
* fixing broken tests