Fix meteor benchmark not running.
This was caused by closure clones not sharing the same rename tables as the base/children
scope.  The code in IRScope.cloneInstrs I think must still be wrong since I do not
think closure scopes should be cloned if the base scope is not clones (else in method).
I will follow up with some hopefully simplification here in a follow up commit.