Always try to unwrap stdio streams.
In order to have stdio streams report their actual file descriptor
(usually 0, 1, 2, but possibly different on Windows), we need to
always try to unwrap the FilterInput/OutputStream that wraps them
on OpenJDK class-library-based JDKs. We do not want to do this
for all processes, since as patched previously some JVMs attempt
to drain subprocess streams. This is not necessary for stdio
streams (as far as we know) so this commit adds a direct unwrap
path we can use when setting up the globals.

Fixes #2598