JavaSupport can't change to interface and be bytecode-compatible.
JavaSupport is exported via Ruby.getJavaSupport as a public API,
and there are extensions that depend on it. By making it an
interface, there's no way to call against it on both 1.7 (where it
is a class) and on 9k (where it is an interface). This commit
changes it back to a class, but makes it purely abstract.