Produce a warning on private attribute accessors
People seem to do that a lot after Sandi's book and not many of them
realize it actually produces a warning in CRuby.

Tried getting TestModule#test_attr out of the CRuby excluded failures,
because it was the only existing test I saw hinting of the behaviour,
however it also test that attribute names can't be a invalid
identifiers, which seems to go fine in JRuby.