Tidy up semantics of matchdata out param in search/match.
Previously the semantics were a bit of a mix; some consumers
passed the out param *and* wanted the matchdata set into context,
some did not want it set, and some did not pass an out param. The
modified logic always uses *either* context.setBackref *or* the
out param in all cases.

Methods that use regexp match/search and need matchdata set now
skip passing the out param, localizing backref setting for those
methods back to where it should live.

This also fixes matchPos to not require the IRubyObject[] out param
if the caller is not interested in the coerced string.