[Truffle] Pulled in almost all of Rubinius's String.
There are a couple bugs in Rubinius 2.4.1's String implementation that are not fixable via monkey-patching.  In these case I removed the lines entirely.  Most of those have been fixed upstream and so future upgrades won't conflict anyway.  The other cases that would fail do so by failing specs, so we'll be able to just upgrade and see what, if anything, is failing.