[Truffle] Fixed String#[](Regexp, int).
When calling String#[] with a Regexp and a negative index, the negative index can't normalize to 0.  Whereas that is valid for MatchData#[], in String#[] it should return nil().  This case in especially annoying because the 'capture' parameter can be any Object that responds to either to_int or to_sym.  Handling all these coercion cases in Java is incredibly verbose, so I opted to execute a Ruby snippet equivalent to what Rubinius would have done.