Rework jar service loading to properly build path.
The logic here will walk from end to beginning of the loading jar
filename, building service class names and attempting to load them.
This is an O(n) operation for the number of path elements, but I
do not know another way to deal with RubyGems canonicalizing jar
file path names.

Conflicts:
	core/src/main/java/org/jruby/runtime/load/ClassExtensionLibrary.java