StaticScope now has Signature instead of three individual fields.  All other changes were made to support this.  We encode/decode scopes with a full Signature.  Blocks are all rooted in providing a Signature.  All IR*Method will provide a Signature.  Unfortunately, I stopped at IR level for method types.  Over time those should get converted.
This corrects all arity issues involving methods and blocks in interp.  JIT
is still somewhere passing in the wrong values but I will correct that in
a followup (this is not a regression).

spec:fast:ruby and test:mri both pass so do JIT variants of those tests
but I strong suspect we are shelling out in some way and running some of
these tests in interp mode.

On other note of distinction is that for builtin functions we only construct
a subset of information that Signature wants because Arity is only a subset
of the info Signature needs.  Since there are no exotic signatures in builtins
there is a Signature.from(Arity) which will convert to a reasonable analogue
or throw if it is special in some way.

I tried to leave all deprecated versions of methods and constructors but
I might have spaced out a few.