[Truffle] Don't override standard execute<type> methods in ToIntNode.
When I added these variants, I didn't want UnexpectedResultException to propagate, so I removed it from the signature.  This changes the way Truffle generates code, however, and it was creating problems for us because the overridden standard methods were still being hooked into the normal specialized node chain, but not propagating the UnexpectedResultException there.  The fix is to use a unique method name in the cases where we don't want UnexpectedResultException to propagate, thereby fixing the standard chain.