Always emit nop for every BB until we can detect truly empty BBs.
The old logic only emitted a nop when a BB was actually empty,
i.e. no instructions at all. However several instructions may be
in the BB but emit no bytecode, like recv_self or check_arity in a
specific-arity method. Because we can't easily detect that a given
BB won't emit any bytecode, we return to unconditionally adding a
nop, so that rescue/trycatch ranges don't end up on empty blocks.

Fixes #2850.