find gem extension classes for uri-like filenames
depending on how the load-service resolves a require ($LOAD_PATH, argument
of require) it can happen that the ClassExtensionLibrary gets a uri-like
searchName. in such a case it needs to be treated as absolute file so the
class-name builder will try the right class-name as well.

fixes #2986

Sponsored By Lookout Inc.