reduce the classloader lookups for require
most creation of  the URLResource are via require 'some/thing' where
it is clear before hand that it is a file and not a directory

URLResource always returns a FileResource, it can be either a file or
a directory or it is marked as not existing. some classloaders return
a stream to directory resource and some classloaders do not. to distinguish
a file from directory you need to check for the .jrubydir file or dig
deeper into the classloader to find out whether this is directory or not.