[Truffle] Ensure local assigns in argument lists are processed before blocks.
If the block references the local var assigned in the argument list before the argument list is processed, it'll end up implicitly declaring a detached local variable with a default value.  When the argument list is processed, the local assign will be at a higher level and declare a local variable with the same name, but it'll be shadowed in the block since the block's local variable will be nearer in scope.

Fixes #2999: Argument values assigned to local variables not available in block.