[Truffle] Upgraded Rubinius core files from 2.4.1 to 2.5.6.
Part of this upgrade was to limit the number of modifications made directly to Rubinius files.  Method overrides are now done in the api/shims files where possible.  Where code cannot be overridden or must otherwise be removed, the Truffle.omit utility was introduce to clearly denote these sections.  The intent is to make it clearer where we can't use Rubinius directly and to make future upgrades go easier.