Convert some very high level execution methods to use RootNode.getFile().
This looks sketchier than it really is.  We know all things returned from parsing will be
a RootNode.  The APIs for the parser existed before RootNode did so it still returns Node.
So any casts being done should be fine.  I also changed one public API method but it really
is an internal method (damn Java packages) for creating an artificial gets loop so the odds
of an extension hooking into this is 0%.