Fixes #3056.  java.lang.IncompatibleClassChangeError in 9.0.0.0 rc1.
This replaces the previous fix.  The previous fix was wrong because the
code in question contained a closure and we were not marking the parent
method as scope as BINDING_HAS_ESCAPED.  As a result, OptimizeDelegationPass
was executing and mucking things up.  The previous fix corrected the pass
to work with nested closures but it would create new problems if that closure
did something which needed to conume the parents binding.

The new fix is just a simple oversight in a few of our instructions.  ZSuper,
UnresolvesSuper, and Match are calls but they were not calling their
superclasses computeScopeFlags.  This is turn made the issue at hand not notice
that the method contained a closure so it never marked BINDING_HAS_ESCAPED.

This fix should potentially fix other weird errors where we did not set common
call IRScope flags.