Fix hanging in modified FIFO read/write spec, tagged in 04db9d8d.
In MRI, rb_read_internal always just does a blind read on the
descriptor, since kernel-level interruption of that blocking call
is safe. On JDK, interruption of an NIO channel read generally
closes that channel, because there's too much state between Java
and the native side to ensure has been cleaned up correctly. Our
normal behavior when dealing with NIO channels is then to always
use select before blocking reads, so we can safely interrupt when
there's no data available. However, this is not appropriate for
native descriptors, like the FIFO in this case, because all ENXIO
channels appear to be selectable (so we attempt it) and the
additional select does not match MRI semantics here.

My change is to modify our version of rb_read_internal to never
select when using a native descriptor. ENXIO should not damage the
descriptor when interrupted, and we should still be able to
interrupt it properly.

Note also that we generally use ENXIO native channels for the
following cases, so only these cases should see any changed
behavior:

* stdio
* some pipes
* most files