Kill unused method invoker logic and related classes.
We have largely committed to JRuby 9k being Java 7+, and to
simplify the process of binding jitted method bodies we are using
MethodHandle exclusively. Because of this, and because MHs work
anywhere reflection works, we are removing the bytecode-generate
CompiledMethod and the Reflected*Method.

There's also a bit of peripheral cleanup for code that is defunct
with these classes removed.