Return always goes to the nearest return-capturing lexical scope.
Return-capturing scopes are:

* method bodies
* lambda blocks
* class and module bodies

Because blocks may be lambdas *or* non-lambdas, we need a live
bit on their dynscope to know it's a block being used as a lambda.
We can't put this bit on the body or the IRScope because they are
static structures.

This fix adds a lambda bit to dynscope and uses that to determine
the nearest return scope. That allows us to eliminate lambda
checks on the return-handling end completely, and always use the
proper scope as our marker for the bubbling return.

Specs coming. Fixes #3143.