CallConfiguration is not used by any DynamicMethods anymore.
CallConfiguration used to be needed to determine how to frame and
scope all methods in the system. However, in JRuby 9k it is no
longer needed.

* Core methods that need frame/scope have the pre/post logic
  generated in bytecode (for generated invokers) or in handles
  (for MethodHandle-based invokers).
* Interpreted IR methods have their own flags for frame/scope, or
  have instructions that perform the pre/post logic.
* Jitted IR methods always have emit bytecode to do the pre/post
  logic.

After a quick inspection I found that there's no other uses of
the callConfig field in DynamicMethod, so this commit removes it
and deprecates the constructors and accessors. This should shrink
all DynamicMethod subclasses by at least 4 bytes and perhaps 8. It
also reduces the size and overhead of our generated populators,
which might have a slight effect on startup.