Move liveness tests out of DCE transform into LVA computation
DCE used to do additional tests to decide if an instruction can be
deleted based on the scope's state. This is a no-no. DCE should
not do any additional tests about deletability based on scope
state (which was in reality liveness tests) -- it should simply use
whatever info is produced by LVA. Those tests should have been
part of LVA instead.

This really didn't affect correctness of DCE, but it would be
providing incorrect information about vars still live when the
scope was exited. Specifically, AddLocalVarLoadStoreInstructions pass
uses that LVA information to determine whether to add lvar stores
on exit. Without accurate LVA information, this can inadvertently
prevent lvar updates from being seen outside the scope.

This is part 1 of the fix for #3173.