Close any open selectors when tearing down a runtime.
This fixes a runtime leak in embedded scenarios where a runtime may
never get garbage collected because it has a thread blocked on a
select call. This leak exists in JRuby 1.7.x as well, but for some
reason I'm seeing it more often in JRuby 9.x when running TorqueBox 3
integration tests. Perhaps there is more IO that uses select in 9.x
than 1.7.x?