Add respond_to-aware hashCode and equals to RubyBasicObject.
We use Object#hashCode and #equals in our Hash impl among other
places. When the object key is RubyObject or a descendant, we
would dispatch dynamically to the appropriate methods in Ruby.
However, RubyBasicObject did not override either method, so they
both just used base JVM impls and ignored user-defined #hash and

This patch adds "checked" versions of these methods to
RubyBasicObject and fixes #3227.