Updates for RSpec 3
Mostly simple syntax changes to the new `expect` style but also various
manual fixes where transpec didn't quite do the job. Normally around
the new `pending` semantics, and that the block form of `pending` is
deprecated.

Another big one was being explicit about the types of errors thrown.

There is still plenty of room for improvement here but at least we are
up and running on the latest/greatest RSpec vintage.