when a container gets terminated then a new container need a new JRubyClassloader
when closing the JRubyClassLoader on a runtime the reference needs to set to null. this
allows to reuse the global runtime which gets a new JRubyClassLoader when used again.
use case is new ScriptionContainer().terminate() then a new instance of ScriptingContainer
has to load jruby extensions as the first instance. note the use of the global runtime here.

fixes #3300