RubyFixnum.equals performs conversion on Float arguments
The 'equals' method is used as an equivalent of `self#==`. `Fixnum#==` converts
Float arguments, so `RubyFixnum.equals` should do so too.

`RubyFixnum.op_equalOther` also converts `Bignum`s which are passed in, so maybe
`RubyFixnum.equals` should do so too. But then again, when can a `Bignum` ever be
equal to a `Fixnum`? In MRI, I'm pretty sure that is impossible, but maybe JRuby
doesn't always "fixize" `Bignum`s?

That is a question for someone else to answer. For now, this fixes a bug.