refactor zoneHelper to better expose tz name logic
In order to reuse and mimic Time#zone behavior the timezone name
logic in RubyTime.zoneHelper needs to be refactored into 2 separated
methods: one that receives the environment TZ as a String and another
that will extract the enviroment TZ from the runtime.

The isTzRelative is only used by RubyTime itself and there is no
need to expose it, thus it was moved back into RubyTime.zone.

Truffle's org.jruby.truffle.noded.TimePrimitiveNodes has been
updated to make use of the new getRubyTimeZoneName method.