Modify blocks to allow jitting like methods.
Lots of duplicated code here between jitting methods and jitting
blocks but the logic is largely the same. MixedModeIRBlockBody,
like the method equivalent, increments a counter until it reaches
JIT threshold. It then forces the block to JIT and binds it into
a CompiledIRBlockBody, which it uses from then on to execute the
block.

Things to do:

* Clean up duplicated code paths and unify more JIT stuff for both
  blocks and methods.
* Reduce indirection through MixedModeIRBlockBody.
* More testing of various forms of blocks.