Synchronize submission of code to JIT.
The old logic could trigger a method or block to JIT twice if two
or more threads all got to that point at the same time. For
methods, this just resulted in wasted work. For blocks, this
appeared to cause some classloading and/or method handle lookup
problems that manifested as NoSuchMethod errors and stack
overflows in the compiled code, likely due to looking up the
wrong method name in the wrong class.

The synchronization probably slows down interpretation, since it
fires for every call when JIT is enabled, so we will want to
change to a lock-free mechanism. This is ok for now, though,
since with JIT enabled both blocks and methods will eventually
stop counting calls and go straight to the jitted body.