Share evalType threadlocal between mixed mode and compiled block.
Fixes #3368

evalType gets tweaked at various stages of block execution, and
when there were two copies in MixedModeIRBlockBody and its
contained CompiledIRBlockBody, they weren't being set consistently
in both places. This change makes them share the threadlocal and
fixes the issues in #3368 leading to eval not scoping right.

The root problem was that evalType is used to indicate that a
given body is being used for an eval, which in this case meant it
was being used for a Class.new class eval. Without the flag, the
search for a method def scope would walk past the Class.new block
and out into the containing class, defining methods in the wrong
place.

TODO: Clean up state like evalType and do a better job of unifying
the different container objects for block bodies.