[Truffle] Fixed a problem with the `string_byte_index` returning the wrong value.
The `string_byte_index` should be returning values normalized to a ByteList 'begin' value of 0. The operations that use the results of this primitive will add the 'begin' value themselves, so this primitive was effectively double-counting the 'begin' value. Of course, when the 'begin' value is 0 this doesn't matter, but in all other cases we end up with invalid values.