Start rooting symbols used as identifiers, to preserve encoding.
The lack of rooting caused a UTF-8 method name in Rails ActionPack
to get lost from the symbol table, and as a result when the list
of methods later turned into a list of symbols, the symbols did
not have proper encoding information.

This is difficult to test, since it will require forcing GC to
run and remove properly-encoded symbols.