Only EPIPE for unseekable streams that are selectable.
We determined some years ago that the only channels/descriptors
for which we should trigger an EPIPE on seek are the same ones
that can be selected against. Reproducing that change here, so
all normal Java streams (InputStream, OutputStream that aren't
convertable into a proper Channel) will silently fail to seek.

Fixes #3399.