Fix GzipWriter output
GZIPOutputStream was passing one of its internal byte buffers to
IOOutputStream#write, which in turn used those bytes to back the
ByteList for its output.  So every subsequent time GZIPOutputStream
wrote to that buffer, it overwrote the bytes behind the back of all
the previous chunks of output.

Ensure that IOOutputStream is working on a copy of the bytes
it gets from GZIPOutputStream to fix this issue.