[Truffle] Truffle::Primitive::Handle generic opaque Java object container.
Useful for when you are writing most of a library in Ruby, with some
primitives in Java and you need to pass objects between those
primitives. Users never see those objects, as you wrap them up in higher
level classes for the user, so you don't want to create a new Ruby class
for each one.