Reimplement Queue and SizedQueue like MRI and add #close.
MRI implements Queue and SizedQueue differently than we had been:

* Implicitly locking around all operations, due to the GIL.
* Notification of just those waiting to push for SizedQueue.

In addition, the new #close method wakes up all threads waiting
to pop or push, which was not easily done by wrapping JDK's
LinkedBlockingQueue.

This new impl is somewhat more in line with MRI, with regards to
locking and notification. This locking will likely introduce some
overhead compared to the old implementation, but the use of a
simple ArrayList may balance that out somewhat. All tests in 2.3's
thread/test_queue.rb pass now except for one relating to a
a peculiarity in MRI's implementation: if you remove marshal_dump
it falls back on Struct marshaling, since in MRI both Queue and
SizedQueue are Structs. Ours are not Struct and I do not see a
good reason to make them be so.